refactoring: addFavorites
