+++ src/shared/hooks/useCheckAuth.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "../../app/providers/store/store";
-import { checkAuth } from "../../features/Authorization";
+import { checkAuth } from "../reducers/Authorization";
 
 export const useAuthCheck = (): void => {
   const dispatch = useDispatch<AppDispatch>();
