+++ src/shared/reducers/Favorites/actions/FavoritesActions.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { arrayRemove, arrayUnion, doc, updateDoc } from "firebase/firestore";
-import { db } from "../../../app/firebase/firebase";
+import { db } from "../../../../firebase";
 
 interface UserFavorites {
   email: string;
