+++ src/shared/reducers/Favorites/index.ts
