+++ src/shared/reducers/Firestor/selectors/FirestorSelectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../../../app/providers/store/store";
+import { RootState } from "../../../../app/providers/store/store";
 
 export const getUser = (state: RootState) => state.firestore.user;
 export const isBdLoading = (state: RootState) => state.firestore.isBdLoading;
