+++ src/shared/reducers/Search/index.ts
